requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
frozenlist==1.2.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
multidict==5.2.0
oauth2client==4.1.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.6
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
